move api testing buttons to components
